fix(nrqlcondition): more granular thresholds
Presently, trying to use a non-integer threshold produces an error in
serializing the API condition, leading to the cryptic log message:

> ERROR controllers.NrqlAlertCondition failed to update condition {"error": "422 response returned: Query can't be blank. Since value must be an integer. Terms must include at least one critical priority term. Terms operator must be one of the following: above, below, equal. Terms priority must be either warning or critical. Terms time function must be one of the following: any, all. Terms duration must be an integer between 1 and 120 (in minutes). Name can't be blank."}

This happens because the resource gets serialized to JSON as a string
with a scale suffix, e.g. "0.5" becomes "500m" (for 500 milli-units).
That doesn't deserialize into a `float64`, as the Go encoder doesn't
know how to parse that format of string.

This change converts the resource.Quantity type into the less-specific
float64 and integer types used by the API types outside of that JSON
encoding flow.