feat: support kubectl diff
In order to permit the API server to call the webhook in "dry run" mode,
it must either announce that it has no side effects or, in the rare case
that it does, knows how to inspect the "dry run" flag and skip those
actions in that mode. The default of "UNKNONWN" means the api-server
fails safe and does not call the webhook.

Since these are simple side-effect free webhooks, we can flag them as
such.