fix: use per-function contexts and printer package
Contexts are meant to be stack-local rather than package-level
containers. The new version(s) of client-go apparently make use of this
context in a way they didn't before, causing nil panics when the context
was not set up properly in tests.

Also handles the move of `NewlineReporter` into a printer package.